+++ onecool_os/connectors/__init__.py
@@ -1,1 +1,9 @@
 """Connector layer package."""
+
+from onecool_os.connectors.import_audit import ImportAudit
+from onecool_os.connectors.import_audit import ImportAuditError
+
+__all__ = [
+    "ImportAudit",
+    "ImportAuditError",
+]